#0: move interleaved addrgen code into separate file
